Cleaned up comments, and restructured code to fix visual bugs

+++ CS3560Shipping/src/shipping/OrderLineItem.java
@@ -66,14 +66,14 @@ public class OrderLineItem
 		return deliverByDate;
 	}
 
-	public double getShippingCost()
+	public BigDecimal getShippingCost()
 	{
-		return shippingCost.doubleValue();
+		return shippingCost;
 	}
 
-	public double getTax()
+	public BigDecimal getTax()
 	{
-		return tax.doubleValue();
+		return tax;
 	}
 
 	public int getQuantity()
